refactor: navigator 관리 상수 및 타입 하나로 합치기
