+++ app/navigation/navigation.ts
@@ -1,8 +1,13 @@
-const navigatorParams = {
+export type NavigatorStackParamList = {
+  [navigatorParams.HOME]: undefined;
+  [navigatorParams.LOGIN]: undefined;
+  [navigatorParams.SIGNUP]: undefined;
+  [navigatorParams.FORGOT_PASSWORD]: undefined;
+};
+
+export const navigatorParams = {
   HOME: 'HOME',
   LOGIN: 'LOGIN',
   SIGNUP: 'SIGNUP',
   FORGOT_PASSWORD: 'FORGOT_PASSWORD',
 } as const;
-
-export default navigatorParams;
